refactor: separate wrapping app for aws lambda handler to dedicated module
